moved geometry classes
