+++ pygeo/__init__.py
@@ -3,11 +3,11 @@ __version__ = "1.6.1"
 from .pyNetwork import pyNetwork
 from .pyGeo import pyGeo
 from .pyBlock import pyBlock
-from .DVGeometryAxi import DVGeometryAxi
 from .constraints.DVCon import DVConstraints
 from .geometry.DVGeo import DVGeometry
+from .geometry.DVGeometryAxi import DVGeometryAxi
 
 try:
-    from .DVGeometryVSP import DVGeometryVSP
+    from .geometry.DVGeometryVSP import DVGeometryVSP
 except ImportError:
     pass
