+++ pygeo/geometry/DVGeometryVSP.py
